+++ webcaldav/db.py
@@ -53,6 +53,8 @@ _USER_SETTINGS_MIGRATION = {
     "double_click_to_create_events": ("BOOLEAN", 0),
     "agenda_search_from_days": ("INTEGER", 0),
     "agenda_search_to_days": ("INTEGER", 365),
+    "holidays_enabled": ("BOOLEAN", 0),
+    "holidays_country": ("TEXT", "'none'"),
 }